Acceptance 리팩토링 1
